+++ src/components/Button/Button.js
@@ -2,10 +2,11 @@ import cn from "classnames";
 
 import s from "./Button.module.scss";
 
-const Button = ({ className, black, children }) => {
+const Button = ({ className, black, children, onClick }) => {
     return (
         <div className={s.root}>
             <button
+                onClick={onClick}
                 className={cn(s.button, className, {
                     [s.colorBlack]: black,
                 })}
