+++ src/components/Button/index.js
@@ -1,3 +1,3 @@
 import Button from "./Button";
 
-export default Button
+export default Button;
